refactor: backport changes to performance tests
* refactor: not existing metrics won't crash app

Changed way of getting some values from metric's map - now get_* functions return some
default values instead of panicking (due to unwrap()), when some key is not existing
in the map.

* refactor: backporting changes to ping test

Backporting changes allowed to increase test ping to 150 while still test time
is reasonable.

* refactor: backporting changes to transaction test

Also, decreased the number of peers to max. 200 as the results above it are
still similar (performance reduction) to what can we observe for iterations
with less peers.

* refactor: backported changes to connection test

* refactor: refactoring done due to clippy suggestions

Some lost suggestions fixed in message and node code along with perf test refactors.

+++ src/protocol/codecs/message.rs
@@ -170,7 +170,7 @@ impl Decoder for MessageCodec {
                 }
 
                 let uncompressed_size = payload_wire_size;
-                let total_wire_size = header_size + payload_wire_size as u32;
+                let total_wire_size = header_size + payload_wire_size;
 
                 let mut message_type = 0;
                 for _ in 0..2 {
